tweaks to tests for handling richtext with html encoding
- separate out the tests
- read-only property to set the regex to identify HTML tags in rich text fields
